requirements.txt
pandas
numpy
scipy
python-dotenv
scikit-learn